feat: 以官方 1.21.1 源码为基线完成移植 + 裁剪 OCCE 附加内容 + 改名 OpenComputersNeo
- 基线切换到 OpenComputers 官方 1.21.1 源码树，mod id 改为 opencomputers_neo
- 全部第三方集成依赖改用本地 libs/deps jar（Modrinth/CurseForge/GitHub 代理获取）
- 裁剪掉相对 1.7.10 原版多出的 19 方块 + 29 物品（Tier4/全息屏/平面屏/磁带机/SSD 等）
- 移除仿 Create:Aeronautics 的创造标签页 banner（mixin + 纹理 + 配置条目）
- 主类改名 OpenComputers -> OpenComputersNeo（611 处引用），配置键/存档目录保持小写 opencomputers
- datagen 生成 blockstates/models（修复紫黑格子），lang/doc/配方/tag 同步裁剪

+++ src/main/scala/li/cil/oc/client/renderer/markdown/segment/TextSegment.scala
@@ -23,9 +23,9 @@ private[markdown] class TextSegment(val parent: Segment, val text: String) exten
       val part = chars.take(numChars)
       hovered = hovered.orElse(resolvedInteractive.fold(None: Option[InteractiveSegment])(_.checkHovered(mouseX, mouseY, currentX, currentY, stringWidth(part, renderer), (Document.lineHeight(renderer) * resolvedScale).toInt)))
       stack.pushPose()
-      stack.translate(currentX, currentY, 0)
+      stack.translate(currentX.toFloat, currentY.toFloat, 0f)
       stack.scale(resolvedScale, resolvedScale, resolvedScale)
-      stack.translate(-currentX, -currentY, 0)
+      stack.translate((-currentX).toFloat, (-currentY).toFloat, 0f)
       graphics.drawString(renderer, resolvedFormat + part, currentX, currentY, resolvedColor)
       stack.popPose()
       currentX = x + wrapIndent
